Extract clip visual geometry into a shared module
The arrangement is moving from an element tree to a painted surface, so
the note preview and clip visuals need one implementation both paths can
read. Living inside the per-clip GPUI element made that impossible
without duplicating it — and two copies of this geometry would drift.

render/clip_geometry.rs is pure: clip data in, paint-ready geometry out,
no GPUI and no theme lookups. midi_clip.rs now renders from it, so the
element path and the coming snapshot painter cannot disagree.

No behavior change; the geometry tests move with the code.
