+++ crates/SphereUIComponents/src/components/timeline/midi_clip.rs
@@ -1,6 +1,11 @@
+use crate::components::timeline::render::clip_geometry::{
+    build_controller_preview, build_note_preview, controller_preview_band_h,
+    midi_controller_default_value, midi_controller_kind_label, visible_clip_px_range,
+    ControllerPreview, NotePreview,
+};
 use crate::components::timeline::timeline_state::{
     midi_debug_enabled, ClipDragItem, ClipEdge, ClipResizeDrag, ClipState, ClipType,
-    MidiControllerKind, MidiControllerPoint, MidiNoteState, TimelineState,
+    MidiControllerKind, TimelineState,
 };
 use crate::theme::Colors;
 use gpui::{
@@ -284,154 +289,6 @@ pub fn midi_clip(
         )
 }
 
-/// Clip-local pixel window that is actually on screen, or `None` when the clip
-/// is fully scrolled out. `left` is the clip's x in lane coordinates.
-fn visible_clip_px_range(left: f32, width: f32, viewport_width: f32) -> Option<(f32, f32)> {
-    let lane_w = viewport_width.max(1.0);
-    let start = (-left).max(0.0);
-    let end = (lane_w - left).min(width);
-    (end > start).then_some((start, end))
-}
-
-/// One horizontal pixel column of coalesced note mass. Pitches are normalized
-/// (0 = bottom of the clip's pitch span, 1 = top) so the paint pass can map them
-/// against the real canvas height without re-reading the notes.
-#[derive(Debug, Clone, Copy)]
-struct NoteColumn {
-    x: f32,
-    lowest_norm: f32,
-    highest_norm: f32,
-}
-
-/// A single note quad, used at zoom levels where notes are individually visible.
-#[derive(Debug, Clone, Copy)]
-struct NoteQuad {
-    x: f32,
-    width: f32,
-    norm_pitch: f32,
-}
-
-/// Resolved note-preview geometry for one clip. Size is bounded by the clip's
-/// visible pixel width, never by its note count.
-struct NotePreview {
-    columns: Vec<NoteColumn>,
-    quads: Vec<NoteQuad>,
-    note_count: usize,
-}
-
-/// Collapse a clip's notes into paintable geometry.
-///
-/// One allocation-free pass over the notes. Raw MIDI pitches are accumulated
-/// into the output slots and normalized afterwards against the clip's full
-/// pitch span, which keeps the whole thing single-pass while still mapping
-/// pitch from the *entire* clip — so the preview does not shift vertically as
-/// the clip scrolls in and out of view.
-fn build_note_preview(
-    notes: &[MidiNoteState],
-    clip_len: f32,
-    ppb: f32,
-    px_start: f32,
-    px_end: f32,
-) -> Option<NotePreview> {
-    if notes.is_empty() || ppb <= 0.0 || px_end <= px_start {
-        return None;
-    }
-
-    let visible_start_beat = (px_start / ppb).max(0.0);
-    let visible_end_beat = (px_end / ppb).min(clip_len).max(0.0);
-    let visible_width = px_end - px_start;
-
-    // Very dense / zoomed-out MIDI maps many notes to the same pixel. Coalesce to
-    // one vertical span per x-column so paint calls stay bounded by clip width
-    // rather than note count while preserving the musical mass. `notes.len()` is
-    // the upper bound on how many land in the window; this is a density heuristic,
-    // so the bound is as good as the exact count and costs no extra pass.
-    let dense = ppb < 5.0 || notes.len() > (visible_width as usize).saturating_mul(3);
-    let columns = visible_width.ceil().clamp(1.0, 2400.0) as usize;
-    let mut spans: Vec<Option<(u8, u8)>> = if dense {
-        vec![None; columns]
-    } else {
-        Vec::new()
-    };
-    let mut raw_quads: Vec<(f32, f32, u8)> = Vec::new();
-    let min_note_w = if ppb < 3.0 { 1.0 } else { 2.0 };
-
-    let mut lo = u8::MAX;
-    let mut hi = 0u8;
-    let mut in_bounds = 0usize;
-    for note in notes {
-        let start = note.start.max(0.0);
-        let end = (note.start + note.duration).min(clip_len);
-        if note.start >= clip_len || note.start + note.duration <= 0.0 || end <= start {
-            continue;
-        }
-        // Pitch span covers the whole clip, not just the visible window.
-        in_bounds += 1;
-        lo = lo.min(note.pitch);
-        hi = hi.max(note.pitch);
-
-        if start >= visible_end_beat || end <= visible_start_beat {
-            continue;
-        }
-        if dense {
-            let x0 = ((start * ppb) - px_start)
-                .floor()
-                .clamp(0.0, (columns - 1) as f32) as usize;
-            let x1 = ((end * ppb) - px_start)
-                .ceil()
-                .clamp(x0 as f32, (columns - 1) as f32) as usize;
-            for cell in &mut spans[x0..=x1] {
-                *cell = Some(match *cell {
-                    Some((low, high)) => (low.min(note.pitch), high.max(note.pitch)),
-                    None => (note.pitch, note.pitch),
-                });
-            }
-        } else {
-            raw_quads.push((
-                start * ppb,
-                ((end - start) * ppb).max(min_note_w),
-                note.pitch,
-            ));
-        }
-    }
-    if in_bounds == 0 {
-        return None;
-    }
-
-    let top_pitch = hi.saturating_add(2).min(127);
-    let bottom_pitch = lo.saturating_sub(2);
-    let pitch_range = (top_pitch as i32 - bottom_pitch as i32).max(12) as f32;
-    let norm_of = |pitch: u8| (pitch as i32 - bottom_pitch as i32) as f32 / pitch_range;
-
-    let columns: Vec<NoteColumn> = spans
-        .into_iter()
-        .enumerate()
-        .filter_map(|(col, span)| {
-            span.map(|(low, high)| NoteColumn {
-                x: px_start + col as f32,
-                lowest_norm: norm_of(low),
-                highest_norm: norm_of(high),
-            })
-        })
-        .collect();
-    let quads: Vec<NoteQuad> = raw_quads
-        .into_iter()
-        .map(|(x, width, pitch)| NoteQuad {
-            x,
-            width,
-            norm_pitch: norm_of(pitch),
-        })
-        .collect();
-    if columns.is_empty() && quads.is_empty() {
-        return None;
-    }
-    Some(NotePreview {
-        columns,
-        quads,
-        note_count: in_bounds,
-    })
-}
-
 fn midi_note_preview_canvas(preview: NotePreview, track_color: gpui::Rgba) -> gpui::Canvas<()> {
     canvas(
         |_bounds, _window, _cx| {},
@@ -474,66 +331,6 @@ fn midi_note_preview_canvas(preview: NotePreview, track_color: gpui::Rgba) -> gp
             }
         },
     )
-}
-
-/// Resolved controller-lane geometry: one normalized value per sampled column,
-/// so the paint pass never touches the (potentially very dense) point lists.
-struct ControllerPreview {
-    lane_kinds: Vec<MidiControllerKind>,
-    /// Per lane, `columns + 1` normalized values sampled left to right.
-    lane_values: Vec<Vec<f32>>,
-    columns: usize,
-    step_px: f32,
-    width: f32,
-}
-
-fn build_controller_preview(
-    lanes: &[crate::components::timeline::timeline_state::MidiControllerLane],
-    clip_len: f32,
-    ppb: f32,
-    width: f32,
-) -> Option<ControllerPreview> {
-    if width <= 1.0 {
-        return None;
-    }
-    let columns = width.ceil().clamp(1.0, 1200.0) as usize;
-    let step_px = (width / columns as f32).max(1.0);
-
-    let mut lane_kinds = Vec::new();
-    let mut lane_values = Vec::new();
-    for lane in lanes
-        .iter()
-        .filter(|lane| lane.visible && !lane.points.is_empty())
-        .take(3)
-    {
-        let default_value = midi_controller_default_value(lane.kind);
-        let mut values = Vec::with_capacity(columns + 1);
-        let mut point_index = 0usize;
-        for col in 0..=columns {
-            let x = (col as f32 * step_px).min(width);
-            let beat = if ppb <= 0.0 {
-                0.0
-            } else {
-                (x / ppb).clamp(0.0, clip_len.max(0.0))
-            };
-            values.push(evaluate_midi_controller_points_cursor(
-                &lane.points,
-                beat,
-                default_value,
-                &mut point_index,
-            ));
-        }
-        lane_kinds.push(lane.kind);
-        lane_values.push(values);
-    }
-
-    (!lane_kinds.is_empty()).then_some(ControllerPreview {
-        lane_kinds,
-        lane_values,
-        columns,
-        step_px,
-        width,
-    })
 }
 
 fn midi_controller_preview_canvas(preview: ControllerPreview) -> gpui::Canvas<()> {
@@ -596,145 +393,4 @@ fn midi_controller_preview_canvas(preview: ControllerPreview) -> gpui::Canvas<()
             }
         },
     )
-}
-
-fn controller_preview_band_h(height: f32, lane_count: usize) -> f32 {
-    let min_needed = (lane_count as f32 * 6.0).max(8.0);
-    (height * 0.44).clamp(min_needed, 30.0).min(height.max(1.0))
-}
-
-fn midi_controller_default_value(kind: MidiControllerKind) -> f32 {
-    match kind {
-        MidiControllerKind::PitchBend => 0.5,
-        MidiControllerKind::CC(_)
-        | MidiControllerKind::ChannelPressure
-        | MidiControllerKind::PolyPressure => 0.0,
-    }
-}
-
-fn evaluate_midi_controller_points_cursor(
-    points: &[MidiControllerPoint],
-    beat: f32,
-    default_value: f32,
-    point_index: &mut usize,
-) -> f32 {
-    if points.is_empty() {
-        return default_value.clamp(0.0, 1.0);
-    }
-    let beat = beat.max(0.0);
-    if beat <= points[0].beat {
-        *point_index = 0;
-        return points[0].value.clamp(0.0, 1.0);
-    }
-    let last = points.len() - 1;
-    if beat >= points[last].beat {
-        *point_index = last.saturating_sub(1);
-        return points[last].value.clamp(0.0, 1.0);
-    }
-
-    while *point_index + 1 < points.len() && beat > points[*point_index + 1].beat {
-        *point_index += 1;
-    }
-    while *point_index > 0 && beat < points[*point_index].beat {
-        *point_index -= 1;
-    }
-    let next = (*point_index + 1).min(last);
-    let a = &points[*point_index];
-    let b = &points[next];
-    let span = (b.beat - a.beat).max(1.0e-6);
-    let t = ((beat - a.beat) / span).clamp(0.0, 1.0);
-    (a.value + (b.value - a.value) * t).clamp(0.0, 1.0)
-}
-
-fn midi_controller_kind_label(kind: MidiControllerKind) -> String {
-    match kind {
-        MidiControllerKind::CC(number) => format!("CC{}", number),
-        MidiControllerKind::PitchBend => "PB".to_string(),
-        MidiControllerKind::ChannelPressure => "AT".to_string(),
-        MidiControllerKind::PolyPressure => "PAT".to_string(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn note(pitch: u8, start: f32, duration: f32) -> MidiNoteState {
-        MidiNoteState::new(pitch, start, duration, 100)
-    }
-
-    #[test]
-    fn offscreen_clips_build_no_preview_geometry() {
-        // Scrolled off to the left, and off to the right of a 1000px lane.
-        assert_eq!(visible_clip_px_range(-500.0, 200.0, 1000.0), None);
-        assert_eq!(visible_clip_px_range(1200.0, 200.0, 1000.0), None);
-    }
-
-    #[test]
-    fn partially_visible_clip_is_clipped_to_the_lane() {
-        // Clip starts 100px left of the lane and runs 400px: only 0..300 shows.
-        let (start, end) = visible_clip_px_range(-100.0, 400.0, 300.0).expect("partially visible");
-        assert_eq!((start, end), (100.0, 400.0));
-        // Clip starts inside the lane and overruns its right edge.
-        let (start, end) = visible_clip_px_range(200.0, 400.0, 300.0).expect("partially visible");
-        assert_eq!((start, end), (0.0, 100.0));
-    }
-
-    #[test]
-    fn dense_preview_stays_bounded_by_visible_pixels_not_note_count() {
-        // 20k notes packed into a clip drawn 200px wide — the pathological
-        // imported-MIDI case that used to emit one quad per note per frame.
-        let notes: Vec<MidiNoteState> = (0..20_000)
-            .map(|i| note(48 + (i % 24) as u8, i as f32 * 0.01, 0.05))
-            .collect();
-        let preview =
-            build_note_preview(&notes, 200.0, 1.0, 0.0, 200.0).expect("notes produce a preview");
-        assert!(
-            preview.quads.is_empty(),
-            "dense zoom coalesces into columns"
-        );
-        assert!(
-            preview.columns.len() <= 200,
-            "columns bounded by visible width, got {}",
-            preview.columns.len()
-        );
-        assert_eq!(preview.note_count, 20_000);
-    }
-
-    #[test]
-    fn zoomed_in_preview_draws_one_quad_per_visible_note() {
-        let notes = vec![note(60, 0.0, 1.0), note(64, 1.0, 1.0), note(67, 2.0, 1.0)];
-        let preview =
-            build_note_preview(&notes, 4.0, 40.0, 0.0, 160.0).expect("notes produce a preview");
-        assert!(preview.columns.is_empty());
-        assert_eq!(preview.quads.len(), 3);
-        assert_eq!(preview.quads[0].width, 40.0);
-    }
-
-    #[test]
-    fn scrolling_culls_notes_without_shifting_the_pitch_mapping() {
-        let notes = vec![note(36, 0.0, 1.0), note(96, 100.0, 1.0)];
-        // The pitch span must come from the whole clip, or the surviving note
-        // would jump vertically as the low note scrolls out of view.
-        let full = build_note_preview(&notes, 200.0, 10.0, 0.0, 2000.0).expect("preview");
-        let scrolled = build_note_preview(&notes, 200.0, 10.0, 990.0, 1020.0).expect("preview");
-        let high_in_full = full
-            .quads
-            .iter()
-            .map(|q| q.norm_pitch)
-            .fold(f32::MIN, f32::max);
-        assert_eq!(scrolled.quads.len(), 1, "only the high note is on screen");
-        assert!(
-            (scrolled.quads[0].norm_pitch - high_in_full).abs() < 1.0e-6,
-            "pitch mapping must not depend on the scroll window"
-        );
-    }
-
-    #[test]
-    fn empty_and_degenerate_inputs_produce_no_preview() {
-        assert!(build_note_preview(&[], 4.0, 40.0, 0.0, 160.0).is_none());
-        // Zero zoom, and a clip whose notes all sit outside its own bounds.
-        assert!(build_note_preview(&[note(60, 0.0, 1.0)], 4.0, 0.0, 0.0, 160.0).is_none());
-        assert!(build_note_preview(&[note(60, 8.0, 1.0)], 4.0, 40.0, 0.0, 160.0).is_none());
-    }
 }
